AccountStatus set up.
-> add dependency org.simplify4u.slf4j-mock.

-> complete AccountStatusServiceImpl.

-> add unit test for AccountStatusServiceImpl.

+++ src/main/java/com/delacasa/auth/repo/AccountStatusRepo.java
@@ -2,14 +2,14 @@ package com.delacasa.auth.repo;
 
 import java.util.List;
 
+import com.delacasa.auth.entity.AccountStatus;
+
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Repository;
 
-import com.delacasa.auth.entity.AccountStatus;
-
 @Repository
 @Scope("prototype")
-public interface AccountStatusRepo extends CustomRepository<AccountStatus, Long> {
+public interface AccountStatusRepo extends CustomRepository<AccountStatus, Integer> {
 
 	List<AccountStatus> findAll();
